Removed program_name from searchable_fields. Renamed RPPA searcher class.
