+++ bq_data_access/v2/feature_search/copynumber_search.py
@@ -51,14 +51,7 @@ class CNVRSearcher(object):
             'label': 'Genomic Build',
             'static': True,
             'values': ['hg19']
-        },
-        {
-            'name': 'program_name',
-            'label': 'Program Name',
-            'static': True,
-            'values': ['tcga']
         }
-
     ]
 
     @classmethod
