+++ bq_data_access/v2/feature_search/gnab_searcher.py
@@ -48,12 +48,6 @@ class GNABSearcher(object):
             'label': 'Genomic Build',
             'static': True,
             'values': ['hg19']
-        },
-        {
-            'name': 'program_name',
-            'label': 'Program Name',
-            'static': True,
-            'values': ['tcga']
         }
     ]
